[refactor] 모임 스케줄 포멧터 로직 수정
